# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Star wars planets API",
    "version": "1.0.1",
    "title": "Star wars planets",
    "contact": {
      "email": "appchto@gmail.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/",
  "tags": [
    {
      "name": "PLANETS",
      "description": "Public calls"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/planet": {
      "get": {
        "tags": [
          "PLANETS"
        ],
        "summary": "Gets planets information",
        "description": "Click Execute button to search for available planets in the system\n",
        "operationId": "searchplanet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "number of records to skip for pagination",
            "required": false,
            "type": "integer",
            "minimum": 0,
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "maximum number of records to return",
            "required": false,
            "type": "integer",
            "maximum": 50,
            "minimum": 0,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "search results matching criteria",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Planets"
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      },
      "post": {
        "tags": [
          "PLANETS"
        ],
        "summary": "adds an planet info",
        "description": "Adds an item to the system",
        "operationId": "addNewPlanet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Planets",
            "description": "planet item to add",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Planets"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "item created"
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        }
      }
    },
    "/planet/{id}": {
      "get": {
        "tags": [
          "PLANETS"
        ],
        "summary": "gets planets info by ID",
        "description": "By passing in the appropriate options, you can search for\navailable planet in the system\n",
        "operationId": "getPlanetWithID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "pass an ID for looking up planet",
            "required": true,
            "type": "string",
            "in": "path"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "maximum number of records to return",
            "required": false,
            "type": "integer",
            "maximum": 50,
            "minimum": 0,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Planets"
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          },
          "default": {
            "Error": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "PLANETS"
        ],
        "summary": "update planet",
        "description": "update planet in the system\n",
        "operationId": "updatePlanet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "pass an ID for looking up the planet",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "Planet",
            "in": "body",
            "description": "planet properties",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Planets"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Planets"
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          },
          "default": {
            "Error": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PLANETS"
        ],
        "summary": "delete a planet by ID",
        "description": "delete a planet in the system\n",
        "operationId": "deletePlanet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "delete a planet",
            "required": true,
            "type": "string",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Planets"
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          },
          "default": {
            "Error": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/planet/{name}": {
      "get": {
        "tags": [
          "PLANETS"
        ],
        "summary": "gets planets info by Name",
        "description": "By passing in the appropriate options, you can search for\navailable planet in the system\n",
        "operationId": "getPlanetWithName",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "description": "pass an name for looking up planet",
            "required": true,
            "type": "string",
            "in": "path"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "maximum number of records to return",
            "required": false,
            "type": "integer",
            "maximum": 50,
            "minimum": 0,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Planets"
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          },
          "default": {
            "Error": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Planets"
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          },
          "default": {
            "Error": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
  },
  "definitions": {
    "Planets": {
      "type": "object",
      "required": [
        "Name",
        "Climate",
        "Terrain"
      ],
      "properties": {
        "Name": {
          "type": "string",
          "example": "Name"
        },
        "Climate": {
          "type": "string",
          "example": "Climate"
        },
        "Terrain": {
          "type": "string",
          "example": "Terrain"
        },
        "Film": {
          "type": "string",
          "example": 0
        },
        "Created_date": {
          "type": "string",
          "format": "date-time",
          "example": "2018-12-28T09:12:33.001Z"
        }
      }
    },
    "GeneralResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "number",
          "description": "returns 1 if successful"
        },
        "description": {
          "type": "string",
          "description": "a short comment"
        },
        "required": [
          "success",
          "description"
        ]
      }
    },
    "ErrorResponse": {
      "required": "message",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}